# src/main/resources/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "contact" : {
      "name" : "PowerCards Support",
      "url" : "https://github.com/victobonetti/powercards"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "title" : "PowerCards API",
    "version" : "1.0.0"
  },
  "paths" : {
...
    "/v1/cards" : {
      "get" : {
        "summary" : "List all cards",
        "parameters" : [ {
          "name" : "deckId",
          "in" : "query",
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 1
          }
        }, {
          "name" : "perPage",
          "in" : "query",
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 20
          }
        }, {
          "name" : "search",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "sort",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PaginatedResponseCardResponse"
                }
              }
            }
          }
        },
        "tags" : [ "Card Resource" ]
      },
      "post" : {
        "summary" : "Create a new card",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CardRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Card created"
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "tags" : [ "Card Resource" ]
      }
    },
    "/v1/cards/bulk/delete" : {
      "post" : {
        "summary" : "Bulk delete cards",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/BulkDeleteRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created"
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "tags" : [ "Card Resource" ]
      }
    },
    "/v1/cards/bulk/move" : {
      "post" : {
        "summary" : "Bulk move cards to deck",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/BulkMoveRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created"
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "tags" : [ "Card Resource" ]
      }
    },
    "/v1/cards/{id}" : {
      "put" : {
        "summary" : "Update an existing card",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CardRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Card updated",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CardResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "Card not found"
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "tags" : [ "Card Resource" ]
      },
      "get" : {
        "summary" : "Get a card by ID",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Card found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CardResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "Card not found"
          }
        },
        "tags" : [ "Card Resource" ]
      },
      "delete" : {
        "summary" : "Delete a card",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Card deleted"
          },
          "404" : {
            "description" : "Card not found"
          }
        },
        "tags" : [ "Card Resource" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "BulkDeleteRequest" : {
        "type" : "object",
        "properties" : {
          "ids" : {
            "type" : "array",
            "items" : {
              "type" : "integer",
              "format" : "int64"
            }
          }
        }
      },
      "BulkMoveRequest" : {
        "type" : "object",
        "properties" : {
          "cardIds" : {
            "type" : "array",
            "items" : {
              "type" : "integer",
              "format" : "int64"
            }
          },
          "targetDeckId" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
...
      "CardRequest" : {
        "type" : "object",
        "properties" : {
          "noteId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "deckId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "ordinal" : {
            "type" : "integer",
            "format" : "int32"
          },
          "modificationTimestamp" : {
            "type" : "integer",
            "format" : "int64"
          },
          "updateSequenceNumber" : {
            "type" : "integer",
            "format" : "int32"
          },
          "type" : {
            "type" : "integer",
            "format" : "int32"
          },
          "queue" : {
            "type" : "integer",
            "format" : "int32"
          },
          "due" : {
            "type" : "integer",
            "format" : "int64"
          },
          "interval" : {
            "type" : "integer",
            "format" : "int32"
          },
          "easeFactor" : {
            "type" : "integer",
            "format" : "int32"
          },
          "repetitions" : {
            "type" : "integer",
            "format" : "int32"
          },
          "lapses" : {
            "type" : "integer",
            "format" : "int32"
          },
          "remainingSteps" : {
            "type" : "integer",
            "format" : "int32"
          },
          "originalDue" : {
            "type" : "integer",
            "format" : "int64"
          },
          "originalDeckId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "flags" : {
            "type" : "integer",
            "format" : "int32"
          },
          "customData" : {
            "type" : "string"
          },
          "noteContent" : {
            "type" : "string"
          },
          "noteTags" : {
            "type" : "string"
          }
        }
      },
      "CardResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "noteId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "deckId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "ordinal" : {
            "type" : "integer",
            "format" : "int32"
          },
          "modificationTimestamp" : {
            "type" : "integer",
            "format" : "int64"
          },
          "updateSequenceNumber" : {
            "type" : "integer",
            "format" : "int32"
          },
          "type" : {
            "type" : "integer",
            "format" : "int32"
          },
          "queue" : {
            "type" : "integer",
            "format" : "int32"
          },
          "due" : {
            "type" : "integer",
            "format" : "int64"
          },
          "interval" : {
            "type" : "integer",
            "format" : "int32"
          },
          "easeFactor" : {
            "type" : "integer",
            "format" : "int32"
          },
          "repetitions" : {
            "type" : "integer",
            "format" : "int32"
          },
          "lapses" : {
            "type" : "integer",
            "format" : "int32"
          },
          "remainingSteps" : {
            "type" : "integer",
            "format" : "int32"
          },
          "originalDue" : {
            "type" : "integer",
            "format" : "int64"
          },
          "originalDeckId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "flags" : {
            "type" : "integer",
            "format" : "int32"
          },
          "customData" : {
            "type" : "string"
          },
          "noteField" : {
            "type" : "string"
          },
          "noteTags" : {
            "type" : "string"
          }
        }
      },
...
      "PaginatedResponseCardResponse" : {
        "type" : "object",
        "properties" : {
          "pagination" : {
            "$ref" : "#/components/schemas/PaginationMeta"
          },
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/CardResponse"
            }
          }
        }
      },
...
      "PaginationMeta" : {
        "type" : "object",
        "properties" : {
          "total" : {
            "type" : "integer",
            "format" : "int64"
          },
          "page" : {
            "type" : "integer",
            "format" : "int32"
          },
          "nextPageUri" : {
            "type" : "string"
          },
          "lastPageUri" : {
            "type" : "string"
          }
        }
      },
...
    }
  },
  "servers" : [ {
    "url" : "http://localhost:8088",
    "description" : "Auto generated value"
  }, {
    "url" : "http://0.0.0.0:8088",
    "description" : "Auto generated value"
  } ]
}